User and staff mutation hooks
