+++ src/staff/views/StaffDetails.tsx
@@ -4,23 +4,9 @@ import NotFoundPage from "@dashboard/components/NotFoundPage";
 import { hasPermissions } from "@dashboard/components/RequirePermissions";
 import { WindowTitle } from "@dashboard/components/WindowTitle";
 import { DEFAULT_INITIAL_SEARCH_DATA } from "@dashboard/config";
-import {
-  PermissionEnum,
-  useChangeStaffPasswordMutation,
-  useStaffAvatarDeleteMutation,
-  useStaffAvatarUpdateMutation,
-  useStaffMemberDeleteMutation,
-  useStaffMemberDetailsQuery,
-  useStaffMemberUpdateMutation,
-} from "@dashboard/graphql";
+import { PermissionEnum, useStaffMemberDetailsQuery } from "@dashboard/graphql";
 import useNavigator from "@dashboard/hooks/useNavigator";
-import useNotifier from "@dashboard/hooks/useNotifier";
-import { commonMessages, errorMessages } from "@dashboard/intl";
-import {
-  extractMutationErrors,
-  getStringOrPlaceholder,
-  maybe,
-} from "@dashboard/misc";
+import { extractMutationErrors, getStringOrPlaceholder } from "@dashboard/misc";
 import usePermissionGroupSearch from "@dashboard/searches/usePermissionGroupSearch";
 import { mapEdgesToItems } from "@dashboard/utils/maps";
 import { DialogContentText } from "@material-ui/core";
@@ -31,6 +17,7 @@ import StaffDetailsPage, {
   StaffDetailsFormData,
 } from "../components/StaffDetailsPage/StaffDetailsPage";
 import StaffPasswordResetDialog from "../components/StaffPasswordResetDialog";
+import { useStaffMutation, useUserMutation } from "../hooks";
 import {
   staffListUrl,
   staffMemberDetailsUrl,
@@ -45,7 +32,6 @@ interface OrderListProps {
 
 export const StaffDetails: React.FC<OrderListProps> = ({ id, params }) => {
   const navigate = useNavigator();
-  const notify = useNotifier();
   const user = useUser();
   const intl = useIntl();
 
@@ -64,23 +50,28 @@ export const StaffDetails: React.FC<OrderListProps> = ({ id, params }) => {
     variables: { id },
     skip: isUserSameAsViewer,
   });
+  const {
+    changePassword,
+    changePasswordOpts,
+    deleteAvatarResult,
+    deleteResult,
+    deleteStaffAvatar,
+    deleteStaffMember,
+    updateStaffAvatar,
+    updateStaffMember,
+    updateStaffMemberOpts,
+  } = useStaffMutation({
+    closeModal,
+    id,
+    refetch,
+  });
+
+  const { updateUserAccount } = useUserMutation();
 
   const staffMember = isUserSameAsViewer ? user.user : data?.user;
   const hasManageStaffPermission = hasPermissions(user.user.userPermissions, [
     PermissionEnum.MANAGE_STAFF,
   ]);
-
-  const [changePassword, changePasswordOpts] = useChangeStaffPasswordMutation({
-    onCompleted: data => {
-      if (data.passwordChange.errors.length === 0) {
-        notify({
-          status: "success",
-          text: intl.formatMessage(commonMessages.savedChanges),
-        });
-        closeModal();
-      }
-    },
-  });
 
   const {
     loadMore: loadMorePermissionGroups,
@@ -91,66 +82,11 @@ export const StaffDetails: React.FC<OrderListProps> = ({ id, params }) => {
     skip: !hasManageStaffPermission,
   });
 
-  const [updateStaffMember, updateStaffMemberOpts] =
-    useStaffMemberUpdateMutation({
-      onCompleted: data => {
-        if (!maybe(() => data.staffUpdate.errors.length !== 0)) {
-          notify({
-            status: "success",
-            text: intl.formatMessage(commonMessages.savedChanges),
-          });
-        }
-      },
-    });
-
-  const [deleteStaffMember, deleteResult] = useStaffMemberDeleteMutation({
-    onCompleted: data => {
-      if (!maybe(() => data.staffDelete.errors.length !== 0)) {
-        notify({
-          status: "success",
-          text: intl.formatMessage(commonMessages.savedChanges),
-        });
-        navigate(staffListUrl());
-      }
-    },
-  });
-
-  const [updateStaffAvatar] = useStaffAvatarUpdateMutation({
-    onCompleted: data => {
-      if (!maybe(() => data.userAvatarUpdate.errors.length !== 0)) {
-        notify({
-          status: "success",
-          text: intl.formatMessage(commonMessages.savedChanges),
-        });
-        refetch();
-      } else {
-        notify({
-          status: "error",
-          title: intl.formatMessage(errorMessages.imgageUploadErrorTitle),
-          text: intl.formatMessage(errorMessages.imageUploadErrorText),
-        });
-      }
-    },
-  });
-
-  const [deleteStaffAvatar, deleteAvatarResult] = useStaffAvatarDeleteMutation({
-    onCompleted: data => {
-      if (!maybe(() => data.userAvatarDelete.errors.length !== 0)) {
-        notify({
-          status: "success",
-          text: intl.formatMessage(commonMessages.savedChanges),
-        });
-        navigate(staffMemberDetailsUrl(id));
-        refetch();
-      }
-    },
-  });
-
   if (staffMember === null) {
     return <NotFoundPage backHref={staffListUrl()} />;
   }
 
-  const handleUpdate = (formData: StaffDetailsFormData) =>
+  const handleStaffUpdate = (formData: StaffDetailsFormData) =>
     extractMutationErrors(
       updateStaffMember({
         variables: {
@@ -163,6 +99,18 @@ export const StaffDetails: React.FC<OrderListProps> = ({ id, params }) => {
             ...(hasManageStaffPermission
               ? groupsDiff(data?.user, formData)
               : {}),
+          },
+        },
+      }),
+    );
+
+  const handleUserUpdate = (formData: StaffDetailsFormData) =>
+    extractMutationErrors(
+      updateUserAccount({
+        variables: {
+          input: {
+            firstName: formData.firstName,
+            lastName: formData.lastName,
           },
         },
       }),
@@ -193,7 +141,7 @@ export const StaffDetails: React.FC<OrderListProps> = ({ id, params }) => {
             }),
           )
         }
-        onSubmit={handleUpdate}
+        onSubmit={isUserSameAsViewer ? handleUserUpdate : handleStaffUpdate}
         onImageUpload={file =>
           updateStaffAvatar({
             variables: {
